删除冗余，修改模块名称
