--- MappingFindViewAnnotation/build.gradle
@@ -1,24 +1,0 @@
-plugins {
-    id 'com.android.library'
-
-}
-apply plugin: 'com.github.dcendents.android-maven'
-group='com.github.zhl45l@163.com'  //注意com.github.github账号
-
-android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.2"
-
-    defaultConfig {
-        minSdkVersion 19
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-}
-
