+++ settings.gradle
@@ -1,3 +1,3 @@
-include ':MappingFindViewAnnotation'
+include ':MappingFindView'
 include ':app'
 rootProject.name = "MappingFindView"
